+++ chain-registry/src/lib.rs
@@ -1,6 +1,8 @@
 pub mod asset_list;
 pub mod chain;
 pub mod error;
+pub mod formatter;
 pub mod paths;
+pub mod querier;
 pub mod relayer_config;
 pub mod utils;
