+++ chain-registry/src/relayer_config.rs
@@ -1,15 +1,13 @@
 //! Contains methods to generate a relayer config for a given chain
-use async_trait::async_trait;
-
 use crate::{
-    asset_list::AssetList, chain::ChainData, error::RegistryError, paths::IBCPath, utils::Fetchable,
+    asset_list::AssetList,
+    chain::ChainData,
+    error::RegistryError,
+    formatter::{GRPCFormatter, UriFormatter},
+    paths::IBCPath,
+    querier::*,
+    utils::Fetchable,
 };
-
-use futures::{stream::FuturesUnordered, StreamExt};
-
-use http::Uri;
-
-use ibc_proto::cosmos::bank::v1beta1::query_client::QueryClient;
 
 use ibc_relayer::{
     config::{
@@ -20,190 +18,10 @@ use ibc_relayer::{
     keyring::Store,
 };
 
-use std::{collections::HashMap, str::FromStr, time::Duration};
+use std::{collections::HashMap, str::FromStr};
 
 use tendermint_light_client_verifier::types::TrustThreshold;
-use tendermint_rpc::{Client, SubscriptionClient, WebSocketClient};
 use tokio;
-use tokio::time::timeout;
-
-// ----------------- RPC ------------------
-
-/// Data retrievable from RPC endpoints.
-#[derive(Clone, Debug)]
-pub struct RpcMandatoryData {
-    pub rpc_address: String,
-    pub max_block_size: u64,
-    pub websocket: tendermint_rpc::Url,
-    // max_block_time should also be retrieved from the RPC
-    // however it looks like it is not in the genesis file anymore
-}
-
-trait QueryInputOutput {
-    type QueryInput : Send;
-    type QueryOutput;
-}
-
-#[async_trait]
-trait QueryContext : QueryInputOutput {
-    fn query_error(chain_name : String) -> RegistryError;
-    async fn query(query: Self::QueryInput) -> Result<Self::QueryOutput, RegistryError>;
-
-    /// Select a healthy RPC/gRPC address from a list of urls.
-    async fn select_healthy(
-        chain_name : String,
-        urls: Vec<Self::QueryInput>,
-    ) -> Result<Self::QueryOutput, RegistryError>
-    {
-        let mut futures: FuturesUnordered<_> = urls.into_iter().map(|url| Self::query(url)).collect();
-
-        while let Some(result) = futures.next().await {
-            if result.is_ok() {
-                return result;
-            }
-        }
-        Err(Self::query_error(chain_name))
-    }
-}
-
-struct RPCQuerier;
-
-impl QueryInputOutput for RPCQuerier {
-    type QueryInput = String;
-    type QueryOutput = RpcMandatoryData;
-}
-
-impl QueryInputOutput for GRPCQuerier {
-    type QueryInput = String;
-    type QueryOutput = tendermint_rpc::Url;
-}
-
-
-#[async_trait]
-impl QueryContext for RPCQuerier {
-    fn query_error(chain_name : String) -> RegistryError{
-        RegistryError::no_healthy_rpc(chain_name)
-    }
-
-    async fn query(rpc: Self::QueryInput) -> Result<Self::QueryOutput, RegistryError> {
-        let websocket_addr = websocket_from_rpc(&rpc)?;
-
-        let (client, driver) = timeout(
-            Duration::from_secs(5),
-            WebSocketClient::new(websocket_addr.clone()),
-        )
-        .await
-        .map_err(|e| RegistryError::websocket_time_out_error(websocket_addr.to_string(), e))?
-        .map_err(|e| RegistryError::websocket_connect_error(websocket_addr.to_string(), e))?;
-
-        let driver_handle = tokio::spawn(async move { driver.run().await });
-
-        let latest_consensus_params = match client.latest_consensus_params().await {
-            Ok(response) => response.consensus_params.block.max_bytes,
-            Err(e) => {
-                return Err(RegistryError::rpc_consensus_params_error(
-                    rpc.to_string(),
-                    e,
-                ))
-            }
-        };
-
-        client
-            .close()
-            .map_err(|e| RegistryError::websocket_conn_close_error(websocket_addr.to_string(), e))?;
-
-        driver_handle
-            .await
-            .map_err(|e| RegistryError::join_error("chain_data_join".to_string(), e))?
-            .map_err(|e| RegistryError::websocket_driver_error(websocket_addr.to_string(), e))?;
-
-        Ok(RpcMandatoryData {
-            rpc_address: rpc.to_string(),
-            max_block_size: latest_consensus_params,
-            websocket: websocket_addr,
-        })
-    }
-}
-
-struct GRPCQuerier;
-
-#[async_trait]
-impl QueryContext for GRPCQuerier {
-    fn query_error(chain_name : String) -> RegistryError{
-        RegistryError::no_healthy_grpc(chain_name)
-    }
-
-    async fn query(grpc: Self::QueryInput) -> Result<Self::QueryOutput, RegistryError> {
-        let uri = parse_or_build_grpc_endpoint(&grpc)?;
-
-        let tendermint_url = uri
-            .to_string()
-            .parse()
-            .map_err(|e| RegistryError::tendermint_url_parse_error(grpc.to_string(), e))?;
-
-        QueryClient::connect(uri)
-            .await
-            .map_err(|_| RegistryError::unable_to_connect_with_grpc())?;
-
-        Ok(tendermint_url)
-    }
-}
-
-
-
-
-// ----------------- websocket ------------------
-
-/// Generates a websocket address from an RPC address.
-fn websocket_from_rpc(rpc_endpoint: &str) -> Result<tendermint_rpc::Url, RegistryError> {
-    let uri = rpc_endpoint
-        .parse::<Uri>()
-        .map_err(|e| RegistryError::uri_parse_error(rpc_endpoint.to_string(), e))?;
-
-    let uri = Uri::builder()
-        .scheme("wss")
-        .authority(
-            uri.authority()
-                .ok_or_else(|| RegistryError::rpc_url_without_authority(rpc_endpoint.to_string()))?
-                .clone(),
-        )
-        .path_and_query("/websocket")
-        .build();
-
-    match uri {
-        Ok(uri) => Ok(tendermint_rpc::Url::from_str(uri.to_string().as_str())
-            .map_err(|e| RegistryError::tendermint_url_parse_error(rpc_endpoint.to_string(), e))?),
-        Err(e) => Err(RegistryError::unable_to_build_websocket_endpoint(
-            rpc_endpoint.to_string(),
-            e,
-        )),
-    }
-}
-
-/// Attempts to parse the given input as a complete URI. If the parsed URI
-/// is not complete, this method attempts to fill in the necessary missing
-/// pieces.
-fn parse_or_build_grpc_endpoint(input: &str) -> Result<Uri, RegistryError> {
-    let uri = input
-        .parse::<Uri>()
-        .map_err(|e| RegistryError::uri_parse_error(input.to_string(), e))?;
-
-    if uri.port().is_none() {
-        return Err(RegistryError::grpc_without_port(input.to_string()));
-    }
-
-    if uri.scheme().is_none() {
-        let builder = Uri::builder();
-        return builder
-            .scheme("https")
-            .authority(input)
-            .path_and_query("/")
-            .build()
-            .map_err(|e| RegistryError::grpc_endpoint_parse_error(input.to_string(), e));
-    }
-
-    Ok(uri)
-}
 
 // ----------------- Packet filters ------------------
 /// Generate packet filters from Vec<IBCPath>.
@@ -243,12 +61,14 @@ async fn hermes_config(
 ) -> Result<ChainConfig, RegistryError> {
     let chain_name = chain_data.chain_name;
 
-    let grpc_endpoints: Vec<String> = chain_data
-        .apis
-        .grpc
-        .iter()
-        .map(|grpc| grpc.address.to_owned())
-        .collect();
+    let mut grpc_endpoints = Vec::new();
+
+    for grpc in chain_data.apis.grpc.iter() {
+        grpc_endpoints.push(GRPCFormatter::parse_or_build_address(
+            grpc.address.as_str(),
+        )?)
+    }
+
     let rpc_endpoints: Vec<String> = chain_data
         .apis
         .rpc
@@ -257,23 +77,12 @@ async fn hermes_config(
         .collect();
 
     let clone_name = chain_name.to_string();
-    let rpc_handle = tokio::spawn(async move {
-        RPCQuerier::select_healthy(
-            clone_name,
-            rpc_endpoints,
-        )
-        .await
-    });
-
+    let rpc_handle =
+        tokio::spawn(async move { RPCQuerier::select_healthy(clone_name, rpc_endpoints).await });
 
     let clone_name = chain_name.to_string();
-    let grpc_handle = tokio::spawn(async move {
-        GRPCQuerier::select_healthy(
-            clone_name,
-            grpc_endpoints,
-        )
-        .await
-    });
+    let grpc_handle =
+        tokio::spawn(async move { GRPCQuerier::select_healthy(clone_name, grpc_endpoints).await });
 
     let base = if let Some(asset) = assets.assets.first() {
         asset.base.clone()
